+++ tests/conftest.py
@@ -1,9 +1,11 @@
 """Test settings."""
 from collections import OrderedDict
+import itertools
 import sys
 import os
 import pytest
 import json
+from textwrap import dedent
 import toml
 import yaml
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
@@ -37,18 +39,31 @@ def mock_settings_prefix(monkeypatch):
     monkeypatch.setattr(settings_parser.Settings, 'prefix', 'TEST_STUFF')
 
 
-@pytest.fixture(params=['.toml', '.yml', '.json'])
-def mock_settings_file(request, monkeypatch, tmpdir):
-    """Temporarily write a settings file and return the filepath and the expected settings outcome."""
-    p = tmpdir.mkdir('sub').join('settings' + request.param)
-    s = """
-    [testgroup]
-    testvar = 123
+@pytest.fixture(params=list(itertools.product(['.toml', '.yml', '.yaml', '.json'], [False, True])))
+def file_extension(request, monkeypatch):
+    """Fixture for providing file extension to use in settings files.
 
-    [othergroup]
-    blabla = 555
+    This fixture is parametrized to mock out all "unneeded" modules to make
+    sure that unneeded libraries do not need to be installed.
+
     """
-    p.write(s)
+    ext, mock_other = request.param
+    if mock_other:
+        if ext == '.toml':
+            monkeypatch.setattr('climatecontrol.settings_parser.yaml', None)
+        elif ext in {'.yml', '.yaml'}:
+            monkeypatch.setattr('climatecontrol.settings_parser.toml', None)
+        else:
+            monkeypatch.setattr('climatecontrol.settings_parser.yaml', None)
+            monkeypatch.setattr('climatecontrol.settings_parser.toml', None)
+    return ext
+
+
+@pytest.fixture
+def mock_settings_file(request, monkeypatch, tmpdir, file_extension):
+    """Temporarily write a settings file and return the filepath and the expected settings outcome."""
+    ext = file_extension
+    p = tmpdir.mkdir('sub').join('settings' + ext)
 
     expected_result = {
         'testgroup': {
@@ -58,43 +73,87 @@ def mock_settings_file(request, monkeypatch, tmpdir):
         }
     }
 
-    if request.param == '.toml':
+    if ext == '.toml':
         p.write(toml.dumps(expected_result))
-    elif request.param == '.yml':
+    elif ext in ['.yml', '.yaml']:
         p.write('---\n' + yaml.dump(expected_result))
-    elif request.param == '.json':
+    elif ext == '.json':
         p.write(json.dumps(expected_result))
+    else:
+        raise NotImplementedError('Invalid file extension :{}.'.format(ext))
 
     return str(p), expected_result
 
 
 @pytest.fixture
-def mock_settings_files(monkeypatch, tmpdir, mock_settings_file):
+def mock_settings_files(request, monkeypatch, tmpdir, mock_settings_file, file_extension):
     """Temporarily write multiple settings file and return the filepaths and the expected settings outcome."""
     subdir = tmpdir.mkdir('sub2')
-    p1 = subdir.join('settings.toml')
-    s = """
-    [testgroup]
-    testvar = 123
+    ext = file_extension
 
-    [othergroup]
-    blabla = 55
-    """
-    p1.write(s)
+    # File to load in settings file by adding "from_file" to the variable we want.
+    inline_path = subdir.join('secret.txt')
+    inline_path.write('foo')
 
-    p2 = subdir.join('settings2.toml')
-    s = """
+    if ext == '.toml':
+        s1 = dedent("""\
+        [testgroup]
+        testvar = 123
+        testvar_inline_1_from_file = "{}"
 
-    [othergroup]
-    blabla = 555
-    """
-    p2.write(s)
+        [othergroup]
+        blabla = 55
+        testvar_inline_2_from_file = "{}"
+        """.format(str(inline_path), str(inline_path)))
+
+        s2 = dedent("""\
+
+        [othergroup]
+        blabla = 555
+        testvar_inline_2 = "bar"
+        """)
+    elif ext in ('.yml', '.yaml'):
+        s1 = dedent("""\
+        testgroup:
+          testvar: 123
+          testvar_inline_1_from_file: {}
+
+        othergroup:
+          blabla: 55
+          testvar_inline_2_from_file: {}
+        """.format(str(inline_path), str(inline_path)))
+
+        s2 = dedent("""\
+
+        othergroup:
+          blabla: 555
+          testvar_inline_2: bar
+        """)
+    elif ext == '.json':
+        s1 = dedent("""\
+        {
+            "testgroup": {"testvar": 123, "testvar_inline_1_from_file": "%s"},
+            "othergroup": {"blabla": 55, "testvar_inline_2_from_file": "%s"}
+        }
+        """) % (str(inline_path), str(inline_path))
+
+        s2 = dedent("""\
+        {"othergroup": {"blabla": 555, "testvar_inline_2": "bar"}}
+        """)
+    else:
+        raise NotImplementedError('Invalid file extension :{}.'.format(ext))
+    p1 = subdir.join('settings' + ext)
+    p1.write(s1)
+    p2 = subdir.join('settings2' + ext)
+    p2.write(s2)
 
     expected_result = {
         'testgroup': {
-            'testvar': 123
+            'testvar': 123,
+            'testvar_inline_1': 'foo'
         }, 'othergroup': {
-            'blabla': 555
+            'blabla': 555,  # Overridden by file 2
+            'testvar_inline_2': 'bar'  # Overridden by file 2
         }
     }
     return [str(p1), str(p2)], expected_result