+++ tests/test_settings.py
@@ -2,6 +2,7 @@
 
 import click
 from click.testing import CliRunner
+import json
 import sys
 import os
 import pytest
@@ -18,6 +19,7 @@ def test_settings_empty(mock_empty_os_environ):
     assert isinstance(settings_map, Mapping)
     assert dict(settings_map) == {}
     assert str(settings_map)  # check that __repr__ works
+    assert len(settings_map) == 0  # length of settings map
 
 
 @pytest.mark.parametrize('update_on_init', [False, True, None])
@@ -80,22 +82,38 @@ def test_parse_from_file_vars(original, file_exists, mock_os_environ, tmpdir):
 @pytest.mark.parametrize('settings_files', ['asd;kjhaflkjhasf', '.', '/home/', ['.', 'asd;kjhaflkjhasf']])
 def test_settings_files_fail(mock_empty_os_environ, settings_files):
     """Check that passing invalid settings files really results in errors."""
-    with pytest.raises(settings_parser.SettingsFileError):
+    with pytest.raises(settings_parser.NoCompatibleLoaderFoundError):
         settings_parser.Settings(prefix='TEST_STUFF',
                                  settings_files='asdlijasdlkjaa')
 
 
 def test_yaml_import_fail(mock_empty_os_environ, monkeypatch):
     """Check that uninstalled yaml really results in an error."""
+    file_str = '---\na: 5'
     # Check that without mocking everything is file:
-    settings_parser.Settings(prefix='TEST_STUFF', settings_files='---\na: 5')
+    settings_parser.Settings(prefix='TEST_STUFF', settings_files=file_str)
     # Now fake not having imported yaml
     monkeypatch.setattr('climatecontrol.settings_parser.yaml', None)
     with pytest.raises(ImportError):
         settings_parser.Settings(prefix='TEST_STUFF', settings_files='---\na: 5')
 
 
-@pytest.mark.parametrize('settings_file_content', ['---\na:\n  b: 5\n', '{"a": {"b": 5}}', '[a]\nb=5'])
+def test_toml_import_fail(mock_empty_os_environ, monkeypatch):
+    """Check that uninstalled toml really results in an error."""
+    file_str = '[section]\na = 5'
+    # Check that without mocking everything is file:
+    settings_parser.Settings(prefix='TEST_STUFF', settings_files=file_str)
+    # Now fake not having imported yaml
+    monkeypatch.setattr('climatecontrol.settings_parser.toml', None)
+    with pytest.raises(ImportError):
+        settings_parser.Settings(prefix='TEST_STUFF', settings_files=file_str)
+
+
+@pytest.mark.parametrize('settings_file_content', [
+    '---\na:\n  b: 5\n',  # yaml
+    '{"a": {"b": 5}}',  # json
+    '[a]\nb=5'  # toml
+])
 def test_settings_file_content(mock_empty_os_environ, settings_file_content):
     """Check parsing file content from different file types works."""
     settings_map = settings_parser.Settings(prefix='TEST_STUFF', settings_files=settings_file_content)
@@ -103,9 +121,9 @@ def test_settings_file_content(mock_empty_os_environ, settings_file_content):
 
 
 @pytest.mark.parametrize('settings_file_content,error', [
-    ('a:\n  b: 5\n', settings_parser.SettingsFileError),  # no file loader with "a" as valid start
+    ('a:\n  b: 5\n', settings_parser.NoCompatibleLoaderFoundError),  # no file loader with "a" as valid start
     ('[{"a": {"b": 5}}]', toml.TomlDecodeError),  # json must be object, seeing "[" assumes a toml file
-    ('b=5', settings_parser.SettingsFileError)  # toml file has to start with [ or it is not parsable
+    ('b=5', settings_parser.NoCompatibleLoaderFoundError)  # toml file has to start with [ or it is not parsable
 ])
 def test_settings_file_content_fail(mock_empty_os_environ, settings_file_content, error):
     """Check parsing file content from different file types raises an error on incorrect file content."""
@@ -113,7 +131,7 @@ def test_settings_file_content_fail(mock_empty_os_environ, settings_file_content
         settings_parser.Settings(prefix='TEST_STUFF', settings_files=settings_file_content)
 
 
-def test_settings_files_file(mock_empty_os_environ, mock_settings_file, tmpdir):
+def test_settings_single_file(mock_empty_os_environ, mock_settings_file, tmpdir):
     """Check that setting a the "settings_files" option works correctly."""
     settings_map = settings_parser.Settings(prefix='TEST_STUFF',
                                             settings_files=mock_settings_file[0])
@@ -121,7 +139,7 @@ def test_settings_files_file(mock_empty_os_environ, mock_settings_file, tmpdir):
     assert dict(settings_map) == mock_settings_file[1]
 
 
-def test_settings_files_files(mock_empty_os_environ, mock_settings_files, tmpdir):
+def test_settings_multiple_files(mock_empty_os_environ, mock_settings_files, tmpdir):
     """Check that setting multiple files as "settings_files" option works correctly."""
     settings_map = settings_parser.Settings(prefix='TEST_STUFF',
                                             settings_files=mock_settings_files[0])
@@ -129,31 +147,10 @@ def test_settings_files_files(mock_empty_os_environ, mock_settings_files, tmpdir
     assert dict(settings_map) == mock_settings_files[1]
 
 
-def test_settings_files_and_env_file(mock_os_environ, mock_settings_files, tmpdir):
-    """Check that using a settings file together with settings parsed from env variables works.
-
-    In the default case environment vars should override settings file vars.
-    """
-    settings_map = settings_parser.Settings(
-        prefix='TEST_STUFF',
-        settings_files=mock_settings_files[0])
-    assert isinstance(settings_map, Mapping)
-    assert dict(settings_map) == {
-        'testgroup': {
-            'test_var': 6,
-            'testvar': 7
-        }, 'othergroup': {
-            'blabla': 555
-        },
-        'testgroup_test_var': 9
-    }
-
-
-def test_settings_files_and_env_file_and_env(mock_env_settings_file, tmpdir):
+def test_settings_env_file_and_env(mock_env_settings_file, tmpdir):
     """Check that a settings file from an env variable works together with other env variables settings.
 
-    For default settings env > env_file > settings file.
-
+    In the default case environment vars should override settings file vars.
     """
     settings_map = settings_parser.Settings(prefix='TEST_STUFF')
     assert isinstance(settings_map, Mapping)
@@ -165,6 +162,64 @@ def test_settings_files_and_env_file_and_env(mock_env_settings_file, tmpdir):
             'blabla': 555
         },
         'testgroup_test_var': 9
+    }
+
+
+def test_settings_multiple_files_and_env(mock_os_environ, mock_settings_files, tmpdir):
+    """Check that using multiple settings files together with settings parsed from env variables works.
+
+    Each subsequent settings file should override the last and environment vars
+    should override any settings file vars.
+
+    """
+    settings_map = settings_parser.Settings(
+        prefix='TEST_STUFF',
+        settings_files=mock_settings_files[0])
+    assert isinstance(settings_map, Mapping)
+
+    assert dict(settings_map) == {
+        'testgroup': {
+            'test_var': 6,
+            'testvar': 7,
+            'testvar_inline_1': 'foo'
+        }, 'othergroup': {
+            'blabla': 555,
+            'testvar_inline_2': 'bar'
+        },
+        'testgroup_test_var': 9
+    }
+
+
+def test_nested_settings_files(tmpdir):
+    """Check that parsing of nested "from_file" settings files works as expected.
+
+    In this case a base file references as settings file (nested_1) which in
+    turn references as second file (nested_2).
+
+    """
+    subfolder = tmpdir.mkdir('sub')
+    p = subfolder.join('settings.json')
+    nested_1_p = subfolder.join('nested_1.json')
+    nested_2_p = subfolder.join('nested_2.json')
+
+    nested_2_p.write(json.dumps({'foo': 1, 'bar': 2}))
+    nested_1_p.write(json.dumps({'level_2_from_file': str(nested_2_p)}))
+    p.write(json.dumps({
+        'level_1_from_file': str(nested_1_p),  # nested_1_p references nested_2_p internally.
+        'spam': 'parrot',
+        'list': [
+            'random',
+            {
+                'this_from_file': str(nested_2_p)  # dictionaries in lists should be expanded as well.
+            }
+        ]
+    }))
+
+    settings_map = settings_parser.Settings(prefix='TEST_STUFF', settings_files=[str(p)])
+    assert dict(settings_map) == {
+        'spam': 'parrot',
+        'level_1': {'level_2': {'foo': 1, 'bar': 2}},
+        'list': ['random', {'this': {'foo': 1, 'bar': 2}}]
     }
 
 
@@ -269,24 +324,23 @@ def test_cli_utils(mock_empty_os_environ, mock_settings_file, mode, option_name,
             '\n').format(option_name, option_name[0])
 
 
-@pytest.mark.parametrize('style', ['.toml', '.json', '.yaml'])
-def test_to_config(mock_empty_os_environ, mock_settings_file, tmpdir, style):
+def test_to_config(mock_empty_os_environ, mock_settings_file, tmpdir, file_extension):
     """Test writing out an example configuration file."""
     settings_file_path, expected = mock_settings_file
     settings_map = settings_parser.Settings(prefix='TEST_STUFF',
                                             settings_files=settings_file_path)
-    s = settings_map.to_config(style=style)
-    if style == '.toml':
+    s = settings_map.to_config(style=file_extension)
+    if file_extension == '.toml':
         expected = mock_settings_file[1]
         assert toml.loads(s) == expected
     else:
         assert s
 
     subdir = tmpdir.mkdir('config_write_subdir')
-    p = subdir.join('example_settings' + style)
+    p = subdir.join('example_settings' + file_extension)
     s = settings_map.to_config(save_to=str(p))
 
-    if style == '.toml':
+    if file_extension == '.toml':
         assert toml.load(str(p)) == expected
     else:
         assert p.read()